codecrafters submit [skip ci]

+++ src/str.cpp
@@ -2,12 +2,12 @@
 
 namespace str {
 std::string ltrim(std::string &s) {
-  int space = s.find(' ');
+  unsigned long space = s.find(' ');
   return s.substr(0, space);
 }
 
 std::string rtrim(std::string &s) {
-  int space = s.find(' ');
+  unsigned long space = s.find(' ');
   return s.substr(space + 1, s.length());
 }
